Move board components to separate directory
