+++ web/client-js/src/components/board/KanbanColumn.vue
